STABLE12-gcc-warnings-cleanup-20030701
FIXES 1650

gcc warnings cleanup; add headers, fix castings


(cherry picked from commit 88e2d8e292e687d3193dcaff62ffd5f2ab4f72ad)

+++ src/afsd/afsd.c
@@ -2088,7 +2088,7 @@ long param1, param2, param3, param4, param5, param6, param7;
     eparm[2] = param6;
     eparm[3] = param7;
 
-    param4 = eparm;
+    param4 = (long) eparm;
 #endif
 
     error = syscall(AFS_SYSCALL, AFSCALL_CALL, param1, param2, param3, param4, param5, param6, param7);
